+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Version 0.5.0
+
+* Php 8.2 in requirements
+* Array doctype from `[]` to `array<>`
+* Test with examples of lib restrictions `\UniMethod\JsonToPhpClass\Tests\Converter\ConverterTest::testBorderRestrictions`
+* `No type` now is `mixed` type
+* Undefined state for properties added
+* Nullable by default in properties
+* Nullable for array is equals '[]'
+* Added Scenarios and Patches, check [ScenariosTest](tests/Builder/Model/ScenariosTest.php)
+
 ## Version 0.2.0
 
 * Camel case for class name